Moving dummy connection and cursor code out to a shared module
