+++ retriever/engines/csvengine.py
@@ -9,25 +9,7 @@ import csv
 from retriever.lib.models import Engine
 from retriever import DATA_DIR, open_fw, open_csvw
 from retriever.lib.tools import sort_csv
-
-
-class DummyConnection(object):
-
-    def cursor(self):
-        pass
-
-    def commit(self):
-        pass
-
-    def rollback(self):
-        pass
-
-    def close(self):
-        pass
-
-
-class DummyCursor(DummyConnection):
-    pass
+from retriever.lib.dummy import DummyConnection, DummyCursor
 
 
 class engine(Engine):
